+++ example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:example/card_example.dart';
+import 'package:example/card_view.dart';
 import 'package:flutter/material.dart';
 import 'package:swipeable_card_stack/swipeable_card_stack.dart';
 
@@ -21,7 +21,6 @@ class MyApp extends StatelessWidget {
 
 class MyHomePage extends StatefulWidget {
   MyHomePage({Key key, this.title}) : super(key: key);
-
   final String title;
 
   @override
@@ -30,18 +29,12 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   int counter = 4;
-  List<CardExample> items = [
-    CardExample(color: Colors.red, text: "First card"),
-    CardExample(color: Colors.blue, text: "Second card"),
-    CardExample(color: Colors.orange, text: "Third card"),
-    CardExample(color: Colors.indigo, text: "Fourth card"),
-    CardExample(color: Colors.green, text: "The next card is the last"),
-    CardExample(color: Colors.purple, text: "This is the last card"),
-  ];
 
   @override
   Widget build(BuildContext context) {
+    //create a CardController
     CardController _cardController = CardController();
+
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
@@ -54,36 +47,29 @@ class _MyHomePageState extends State<MyHomePage> {
             context: context,
             //add the first 3 cards
             items: [
-              CardExample(color: Colors.red, text: "First card"),
-              CardExample(color: Colors.blue, text: "Second card"),
-              CardExample(color: Colors.orange, text: "Third card"),
+              CardView(text: "First card"),
+              CardView(text: "Second card"),
+              CardView(text: "Third card"),
             ],
-            onCardSwiped: (dir) {
-              //Add the next item if available
+            onCardSwiped: (dir, index, widget) {
+              //Add the next card
               if (counter <= 20) {
-                _cardController.addItem(CardExample(
-                    color: Colors.greenAccent, text: "card $counter"));
+                _cardController.addItem(CardView(text: "Card $counter"));
                 counter++;
               }
 
               if (dir == Direction.left) {
-                print('onDisliked');
+                print('onDisliked ${(widget as CardView).text} $index');
               } else if (dir == Direction.right) {
-                print('onLiked');
+                print('onLiked ${(widget as CardView).text} $index');
               } else if (dir == Direction.up) {
-                print('onUp');
+                print('onUp ${(widget as CardView).text} $index');
               } else if (dir == Direction.down) {
-                print('onDown');
+                print('onDown ${(widget as CardView).text} $index');
               }
-            },
-            onLastCardSwiped: () {
-              print('onLastCardSwiped');
             },
             enableSwipeUp: true,
             enableSwipeDown: true,
-            // cardHeightTopMul: 0.3,
-            // cardHeightMiddleMul: 0.25,
-            // cardHeightBottomMul: 0.2,
           ),
           Container(
             margin: EdgeInsets.symmetric(vertical: 20.0),
@@ -91,20 +77,21 @@ class _MyHomePageState extends State<MyHomePage> {
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: [
                 FloatingActionButton(
-                    child: Icon(Icons.chevron_left),
-                    onPressed: () => _cardController.triggerSwipeLeft()),
+                  child: Icon(Icons.chevron_left),
+                  onPressed: () => _cardController.triggerSwipeLeft(),
+                ),
                 FloatingActionButton(
-                    child: Icon(Icons.chevron_right),
-                    onPressed: () => _cardController.triggerSwipeRight()),
+                  child: Icon(Icons.chevron_right),
+                  onPressed: () => _cardController.triggerSwipeRight(),
+                ),
                 FloatingActionButton(
-                    child: Icon(Icons.arrow_upward),
-                    onPressed: () => _cardController.triggerSwipeUp()),
+                  child: Icon(Icons.arrow_upward),
+                  onPressed: () => _cardController.triggerSwipeUp(),
+                ),
                 FloatingActionButton(
-                    child: Icon(Icons.arrow_downward),
-                    onPressed: () => _cardController.triggerSwipeDown()),
-                // FloatingActionButton(
-                //     child: Icon(Icons.replay_outlined),
-                //     onPressed: () => _cardController.appendItems(moreItems))
+                  child: Icon(Icons.arrow_downward),
+                  onPressed: () => _cardController.triggerSwipeDown(),
+                ),
               ],
             ),
           )