+++ lib/swipe_controller.dart
@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
 
 typedef TriggerListener = void Function(Direction dir);
-// typedef AppendItems = void Function(List<CardItem> moreItems);
 typedef AppendItem = void Function(Widget item);
 typedef EnableSwipe = void Function(bool dir);
 
 class CardController {
   TriggerListener listener;
-  // AppendItems addItems;
   AppendItem addItem;
   EnableSwipe enableSwipeListener;
 
@@ -26,10 +24,6 @@ class CardController {
   void triggerSwipeDown() {
     return listener.call(Direction.down);
   }
-
-  // void appendItems(List<CardItem> moreItems) {
-  //   return addItems.call(moreItems);
-  // }
 
   void appendItem(Widget item) {
     return addItem.call(item);