+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - February 19, 2021
 
-* TODO: Describe initial release.
+- Added `SwipeableCardsSection`
+  - Can add custom widgets to SwipeableCardsSection
+  - Cards can be swiped horizontally and vertically
+  - Can receive swipe events
+- Added `CardController` to automatically swipe cards:
+  - `_cardController.triggerSwipeLeft()`
+  - `_cardController.triggerSwipeRight()`
+  - `_cardController.triggerSwipeUp()`
+  - `_cardController.triggerSwipeDown()`
+- Added example (see `examples/` folder)